requirements.txt
numpy==1.21.2
opencv-python==4.5.3.56
imagesize
ExifRead
tqdm
PyYAML==5.4.1